backend/requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
python-dotenv
pydantic

langchain
langchain-groq
